components: move SFXBank ctx-resource owner beside sfx_bank_system (#1358)
Follow-up to #1351 / PR #1357: SFXBank is the same family of ctx-singleton
audio resource owner as MusicContext / TextContext / RenderTargets /
camera::ShapeMeshes were before #1351. It owns raylib Sound handles, has
deleted copy, custom move ctor/assign, a destructor, and a release() that
calls UnloadSound — i.e. RAII, not entity-owned plain data.

Per .squad/decisions.md §"app/components parallel audit verdict
(consolidated)":

  > app/components/ stays reserved for atomic, queryable entity-owned data.
  > Dispatcher payloads, ctx scratch buffers, session state, persistence
  > helpers, and render-config helpers should move beside their owning
  > systems/modules.

Changes:

- New app/systems/sfx_bank_resources.h declares the SFXBank struct (members,
  deleted copy, move ctor/assign, dtor, release()). Definitions stay in
  app/systems/sfx_bank.cpp.
- app/components/audio.h drops the SFXBank struct and the <raylib.h>
  include; keeps only the SFX enum + contiguity static_asserts. The header
  is now trivially destructible plain data.
- sfx_bank.cpp / audio_system.cpp / test_audio_system.cpp /
  test_gpu_resource_lifecycle.cpp pull in the new
  systems/sfx_bank_resources.h for the SFXBank definition; they still
  include components/audio.h for the SFX enum.
- Drop now-unused components/audio.h includes from app/game_loop.cpp and
  benchmarks/bench_systems.cpp (both only used sfx_bank_init /
  wire_audio_haptic_dispatcher, not SFX/SFXBank directly).

Closes #1358.

Build green with zero warnings (-Wall -Wextra -Werror), 959 test cases /
3357 assertions pass.
